// web/src/shared/api/generated/model/examBlueprintInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ExamDifficultyMixInput } from './examDifficultyMixInput';
import type { ExamTypeCountInput } from './examTypeCountInput';

export interface ExamBlueprintInput {
  typeCounts: ExamTypeCountInput[];
  difficultyMix: null | ExamDifficultyMixInput;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  timeLimitMinutes: number | string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  passMark: number | string;
}