// web/src/shared/api/generated/model/unitExamBlueprintResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ExamBlueprintResult } from './examBlueprintResult';
import type { ExamTypeCountResult } from './examTypeCountResult';

export interface UnitExamBlueprintResult {
  unitId: string;
  name: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  order: number | string;
  blueprint: null | ExamBlueprintResult;
  servable: ExamTypeCountResult[];
}